# docs/openapi-gpt-3.0.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "esdata API",
    "description": "API de consulta fiscal-regulatoria con trazabilidad a fuente oficial, superficies de governance y workers observables via sync_log.",
    "version": "0.1.6"
  },
  "servers": [
    {
      "url": "https://api.desuscribir.es",
      "description": "Production API"
    }
  ],
  "paths": {
...
    "/v1/irs-fiscal/normas": {
      "get": {
        "tags": [
          "irs-fiscal"
        ],
        "summary": "Listar Normas Irs",
        "operationId": "listar_normas_irs",
        "parameters": [
          {
            "name": "tipo",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "Filtrar por tipo: publicacion, forma, instruccion, ley, convenio",
              "title": "Tipo"
            },
            "description": "Filtrar por tipo: publicacion, forma, instruccion, ley, convenio"
          },
          {
            "name": "estado",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "Estado: activo, inactivo, obsoleto",
              "default": "activo",
              "title": "Estado"
            },
            "description": "Estado: activo, inactivo, obsoleto"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IrsFiscalNormaListResponse"
                }
              }
            }
          }
        }
      }
    },
    "/v1/irs-fiscal/w8-formularios": {
      "get": {
        "tags": [
          "irs-fiscal"
        ],
        "summary": "Listar Formularios W8",
        "operationId": "listar_formularios_w8",
        "parameters": [
          {
            "name": "tipo_sujeto",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "Filtrar por tipo: persona_fisica, persona_juridica, exento",
              "title": "Tipo Sujeto"
            },
            "description": "Filtrar por tipo: persona_fisica, persona_juridica, exento"
          },
          {
            "name": "estado",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "Estado: activo, inactivo",
              "default": "activo",
              "title": "Estado"
            },
            "description": "Estado: activo, inactivo"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IrsW8FormListResponse"
                }
              }
            }
          }
        }
      }
    },
    "/v1/irs-fiscal/tin-referencias": {
      "get": {
        "tags": [
          "irs-fiscal"
        ],
        "summary": "Listar Referencias Tin",
        "operationId": "listar_referencias_tin",
        "parameters": [
          {
            "name": "pais",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "Filtrar por codigo de pais (ISO 3166-1 alpha-2)",
              "title": "Pais"
            },
            "description": "Filtrar por codigo de pais (ISO 3166-1 alpha-2)"
          },
          {
            "name": "ocde",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "description": "Filtrar por pais OCDE",
              "title": "Ocde"
            },
            "description": "Filtrar por pais OCDE"
          },
          {
            "name": "eu_vat",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "description": "Filtrar por pais UE VAT",
              "title": "Eu Vat"
            },
            "description": "Filtrar por pais UE VAT"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IrsTinReferenceListResponse"
                }
              }
            }
          }
        }
      }
    },
    "/v1/irs-fiscal/giin": {
      "get": {
        "tags": [
          "irs-fiscal"
        ],
        "summary": "Listar Registros Giin",
        "operationId": "listar_registros_giin",
        "parameters": [
          {
            "name": "estado",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "Filtrar por estado FATCA: activo, inactivo, suspendido",
              "title": "Estado"
            },
            "description": "Filtrar por estado FATCA: activo, inactivo, suspendido"
          },
          {
            "name": "pais",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "Filtrar por pais de la entidad",
              "title": "Pais"
            },
            "description": "Filtrar por pais de la entidad"
          },
          {
            "name": "tipo",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "Filtrar por tipo: FFI, NFFE, Exempt Beneficial Owner",
              "title": "Tipo"
            },
            "description": "Filtrar por tipo: FFI, NFFE, Exempt Beneficial Owner"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 500,
              "minimum": 1,
              "description": "Limite de registros devueltos",
              "default": 50,
              "title": "Limit"
            },
            "description": "Limite de registros devueltos"
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "description": "Offset de paginacion",
              "default": 0,
              "title": "Offset"
            },
            "description": "Offset de paginacion"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GiinRegistryListResponse"
                }
              }
            }
          }
        }
      }
    },
    "/v1/irs-fiscal/giin/{giin}": {
      "get": {
        "tags": [
          "irs-fiscal"
        ],
        "summary": "Detalle Registro Giin",
        "operationId": "detalle_registro_giin",
        "parameters": [
          {
            "name": "giin",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Giin"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GiinRegistryDetail"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "security": [
    {
      "ApiKeyAuth": []
    }
  ],
  "components": {
    "schemas": {
...
      "GiinRegistryDetail": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "giin": {
            "type": "string",
            "title": "Giin"
          },
          "entidad_nombre": {
            "type": "string",
            "title": "Entidad Nombre"
          },
          "entidad_pais": {
            "type": "string",
            "title": "Entidad Pais"
          },
          "tipo_entidad": {
            "type": "string",
            "title": "Tipo Entidad"
          },
          "estado_fatca": {
            "type": "string",
            "title": "Estado Fatca"
          },
          "fecha_expiracion": {
            "type": "string",
            "title": "Fecha Expiracion"
          },
          "fecha_registro": {
            "type": "string",
            "title": "Fecha Registro"
          },
          "es_exempt_beneficial_owner": {
            "type": "boolean",
            "title": "Es Exempt Beneficial Owner"
          },
          "es_sponsored_ffo": {
            "type": "boolean",
            "title": "Es Sponsored Ffo"
          },
          "nota": {
            "type": "string",
            "title": "Nota"
          },
          "creado_en": {
            "type": "string",
            "title": "Creado En"
          },
          "actualizado_en": {
            "type": "string",
            "title": "Actualizado En"
          }
        },
        "type": "object",
        "required": [
          "id",
          "giin",
          "entidad_nombre",
          "tipo_entidad",
          "estado_fatca"
        ],
        "title": "GiinRegistryDetail"
      },
      "GiinRegistryListResponse": {
        "properties": {
          "registros": {
            "items": {
              "$ref": "#/components/schemas/GiinRegistrySummary"
            },
            "type": "array",
            "title": "Registros"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "limit": {
            "type": "integer",
            "title": "Limit"
          },
          "offset": {
            "type": "integer",
            "title": "Offset"
          },
          "has_more": {
            "type": "boolean",
            "title": "Has More"
          },
          "next_offset": {
            "type": "integer",
            "title": "Next Offset"
          }
        },
        "type": "object",
        "required": [
          "total"
        ],
        "title": "GiinRegistryListResponse"
      },
      "GiinRegistrySummary": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "giin": {
            "type": "string",
            "title": "Giin"
          },
          "entidad_nombre": {
            "type": "string",
            "title": "Entidad Nombre"
          },
          "entidad_pais": {
            "type": "string",
            "title": "Entidad Pais"
          },
          "tipo_entidad": {
            "type": "string",
            "title": "Tipo Entidad"
          },
          "estado_fatca": {
            "type": "string",
            "title": "Estado Fatca"
          },
          "fecha_expiracion": {
            "type": "string",
            "title": "Fecha Expiracion"
          }
        },
        "type": "object",
        "required": [
          "id",
          "giin",
          "entidad_nombre",
          "tipo_entidad",
          "estado_fatca"
        ],
        "title": "GiinRegistrySummary"
      },
...
      "IrsFiscalNormaListResponse": {
        "properties": {
          "normas": {
            "items": {
              "$ref": "#/components/schemas/IrsFiscalNormaSummary"
            },
            "type": "array",
            "title": "Normas"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          }
        },
        "type": "object",
        "required": [
          "total"
        ],
        "title": "IrsFiscalNormaListResponse"
      },
      "IrsFiscalNormaSummary": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "codigo": {
            "type": "string",
            "title": "Codigo"
          },
          "titulo": {
            "type": "string",
            "title": "Titulo"
          },
          "tipo": {
            "type": "string",
            "title": "Tipo"
          },
          "anio_vigencia": {
            "type": "integer",
            "title": "Anio Vigencia"
          },
          "estado": {
            "type": "string",
            "title": "Estado"
          }
        },
        "type": "object",
        "required": [
          "id",
          "codigo",
          "titulo",
          "tipo",
          "estado"
        ],
        "title": "IrsFiscalNormaSummary"
      },
      "IrsTinReferenceListResponse": {
        "properties": {
          "referencias": {
            "items": {
              "$ref": "#/components/schemas/IrsTinReferenceSummary"
            },
            "type": "array",
            "title": "Referencias"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          }
        },
        "type": "object",
        "required": [
          "total"
        ],
        "title": "IrsTinReferenceListResponse"
      },
      "IrsTinReferenceSummary": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "codigo_pais": {
            "type": "string",
            "title": "Codigo Pais"
          },
          "pais_nombre": {
            "type": "string",
            "title": "Pais Nombre"
          },
          "formato_tin": {
            "type": "string",
            "title": "Formato Tin"
          },
          "ejemplo_tin": {
            "type": "string",
            "title": "Ejemplo Tin"
          },
          "es_ocde": {
            "type": "boolean",
            "title": "Es Ocde"
          },
          "es_eu_vat": {
            "type": "boolean",
            "title": "Es Eu Vat"
          }
        },
        "type": "object",
        "required": [
          "id",
          "codigo_pais",
          "pais_nombre",
          "es_ocde",
          "es_eu_vat"
        ],
        "title": "IrsTinReferenceSummary"
      },
      "IrsW8FormListResponse": {
        "properties": {
          "formularios": {
            "items": {
              "$ref": "#/components/schemas/IrsW8FormSummary"
            },
            "type": "array",
            "title": "Formularios"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          }
        },
        "type": "object",
        "required": [
          "total"
        ],
        "title": "IrsW8FormListResponse"
      },
      "IrsW8FormSummary": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "codigo": {
            "type": "string",
            "title": "Codigo"
          },
          "nombre": {
            "type": "string",
            "title": "Nombre"
          },
          "tipo_sujeto": {
            "type": "string",
            "title": "Tipo Sujeto"
          },
          "validez_anios": {
            "type": "integer",
            "title": "Validez Anios"
          },
          "estado": {
            "type": "string",
            "title": "Estado"
          }
        },
        "type": "object",
        "required": [
          "id",
          "codigo",
          "nombre",
          "tipo_sujeto",
          "estado"
        ],
        "title": "IrsW8FormSummary"
      },
...
    },
    "securitySchemes": {
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key",
        "description": "Dedicated ESData API key for GPT Actions."
      }
    }
  }
}